+++ RELEASES.md
@@ -1,6 +1,7 @@
 Version 0.9.7 (2024-XX-XX)
 ========================
 * New driving model
+* Hyundai Elantra 2024 support
 
 Version 0.9.6 (2024-02-27)
 ========================